backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Async and HTTP
aiohttp==3.9.1
httpx==0.26.0
websockets==12.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Technical Analysis
TA-Lib==0.4.28
talib-binary==0.4.28  # Precompiled for easy installation

# Caching
cachetools==5.3.2

# Database (optional for future)
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
